requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
qdrant-client==1.7.3
redis==5.0.1
python-dotenv==1.0.1
httpx==0.27.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
pydantic==2.6.1